Fix flake8 issues, move TracingWrappingConnection to separate package
